+++ web/middleware.go
@@ -7,6 +7,13 @@ import (
 	"github.com/valuechaintool/valuechaintool/models"
 )
 
+func MiddlewareHSTS() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.Writer.Header().Set("Strict-Transport-Security", "max-age=63072000")
+		c.Next()
+	}
+}
+
 func MiddlewareAuth() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		token, err := c.Cookie("session_id")