+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## [Unreleased]
 ### Features
 - Allow to expose HTTPS
+- Implement HSTS header
 
 <a name="0.2.0"></a>
 ## [0.2.0] - 2020-06-05